--- src/components/Login/empty.ts
@@ -1,1 +1,0 @@
-// 실제 사용하실 때 본 파일은 삭제 후 사용하시면 됩니다.
